AI Praticioner Portal/backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pytest>=7.4.3
pytest-cov>=4.1.0
passlib>=1.7.4
cryptography>=41.0.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
